feat: refactor version display into ReleasedVersions component
